ready all except rating bar
