+++ app/src/main/java/com/kamikadze328/mtstetaproject/data/dto/Actor.kt
@@ -1,8 +1,8 @@
-package com.kamikadze328.mtstetaproject.objects
+package com.kamikadze328.mtstetaproject.data.dto
 
 import android.graphics.drawable.Drawable
 
-data class Actor(
+data class ActorDto(
     val avatarIcon: Drawable?,
     val name: String
 )
