+++ app/src/main/res/layout/item_actor.xml
